+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Change Log
 ===============================================================================
 
+Version 1.X.X *(2014-XX-XX)*
+----------------------------
+* LIB-EXTRA: Added a new method to add external adapter different from CardGridStaggeredArrayAdapter  [(doc)](https://github.com/gabrielemariotti/cardslib/blob/master/doc/STAGGEREDGRID.md#how-to-use-an-external-adapter)
+
+
 Version 1.5.0 *(2014-03-21)*
 ----------------------------
 * LIB-EXTRA: Added a `CardGridStaggeredView` [(doc)](https://github.com/gabrielemariotti/cardslib/blob/master/doc/STAGGEREDGRID.md)